+++ lib/berkshelf/extensions.rb
@@ -1,0 +1,1 @@
+require 'berkshelf/extensions/cached_cookbook'