+++ lib/strainer/color.rb
@@ -1,7 +1,0 @@
-require 'term/ansicolor'
-
-module Strainer
-  class Color
-    extend Term::ANSIColor
-  end
-end