+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Strainer CHANGELOG
 ==================
 
+v1.0.0
+------
+- Moved entirely to Berkshelf integration
+- Moved entirely to thor
+- **Breaking** - new command `strainer`, old command is deprecated
+
 v0.2.1
 ------
 - Support a wider range of chef versions